requirements.txt
annotated-types==0.7.0
asn1crypto==1.5.1
boto3==1.40.41
botocore==1.40.41
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
cryptography==46.0.0
Faker==28.4.1
filelock==3.19.1
idna==3.10
jmespath==1.0.1
optional-faker==2.1.0
packaging==25.0
platformdirs==4.4.0
pycparser==2.23
pydantic==2.11.9
pydantic_core==2.33.2
PyJWT==2.10.1
pyOpenSSL==25.3.0
python-dateutil==2.9.0.post0
python-dotenv==0.21.0
python-rapidjson==1.5
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
s3transfer==0.14.0
six==1.17.0
snowflake==1.8.0
snowflake-connector-python==3.17.4
snowflake._legacy==1.0.2
snowflake.core==1.8.0
sortedcontainers==2.4.0
tomlkit==0.13.3
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
